\name{Bassler2004}
\docType{data}
\alias{Bassler2004}

\title{Studies on Ketotifen Alone or as Additional Medication for Long-Term Control of Asthma and Wheeze in Children}

\description{
Results from ten trials reporting the physicians' judgement on the overall
efficacy of ketotifen for long-term control of asthma and wheeze in children.
}

\usage{Bassler2004}

\format{
The data frame contains the following columns:
\tabular{lll}{
\bold{study} \tab \code{character} \tab study label \cr
\bold{Ee}    \tab \code{integer}   \tab number of children with treament success (ketotifen group) \cr
\bold{Ne}    \tab \code{integer}   \tab number of children (ketotifen group) \cr
\bold{Ec}    \tab \code{integer}   \tab number of children with treament success (control group) \cr
\bold{Nc}    \tab \code{integer}   \tab number of children (control group) \cr
\bold{blind} \tab \code{character} \tab blinding of clinicians \cr
}
}

\details{
Results from ten trials reporting the physicians' judgement on the overall
efficacy of Ketotifen for long-term control of asthma and wheeze in children.
A prespecified subgroup analysis was conducted to evaluate whether the
treatment effect is different in trials with adequate blinding compared to
trials with inadequate / unclear blinding.

This data set is used as an example in Schwarzer et al. (2015).
}

\source{
Bassler D, Mitra AAD, Ducharme FM, Forster J, Schwarzer G (2004):
Ketotifen alone or as additional medication for long-term control of asthma and
wheeze in children.
\emph{Cochrane Database of Systematic Reviews},
\bold{1}, CD001384.
\doi{10.1002/14651858.CD001384.pub2}}

\references{
Schwarzer G, Carpenter JR and Rücker G (2015):
\emph{Meta-Analysis with R (Use-R!)}.
Springer International Publishing, Switzerland
}

\author{Guido Schwarzer, \email{guido.schwarzer@uniklinik-freiburg.de}}

\examples{
\donttest{
if (requireNamespace("meta", quietly = TRUE)) {
# Load meta package
library("meta")

# Use DerSimonian-Laird estimator (which was the default in meta in the year 2015).
# Furthermore, print meta-analysis results with two digits.
oldset <- settings.meta(method.tau = "DL", digits = 2)

# Calculate experimental and control event rates
print(with(Bassler2004, summary(Ee / Ne)))
print(with(Bassler2004, summary(Ec / Nc)))

# Conduct meta-analysis using the inverse variance method
ma <- metabin(Ee, Ne, Ec, Nc, method = "I",
  data = Bassler2004, studlab = study)
print(ma)

# Conduct subgroup analysis comparing trials with adequate blinding
# to trials with inadequate or unclear blinding
ma_blind <- update(ma, subgroup = blind, print.subgroup.name = FALSE)
print(ma_blind)

# Conduct subgroup analysis assuming common between-study variance in subgroups
ma_blind_c <- update(ma_blind, tau.common = TRUE)
print(ma_blind_c)

# Use previous settings
settings.meta(oldset)
}
}
}

\keyword{data sets}
\concept{risk ratios}
\concept{medicine}
\concept{subgroup analysis}
\section{Concepts}{
   risk ratios, medicine, subgroup analysis
}
