% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabook-package.R
\docType{package}
\name{metabook-package}
\alias{metabook-package}
\alias{metabook}
\title{metabook: Data Sets and Code for "Meta-Analysis with R"}
\description{
Data sets and code supporting the second edition of "Meta-Analysis with R"
by Schwarzer, Carpenter, and Rücker.
}
\references{
Balduzzi S, Rücker G, Schwarzer G (2019):
How to perform a meta-analysis with R: a practical tutorial.
\emph{Evidence-Based Mental Health},
\bold{22}, 153--160

Schwarzer G, Carpenter JR and Rücker G (2015):
\emph{Meta-Analysis with R (Use-R!)}.
Springer International Publishing, Switzerland
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/guido-s/metabook/}
}

}
\author{
Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
\keyword{package}
